+++ versions/1.18.2/src/main/java/com/ayakacraft/carpetayakaaddition/CarpetAyakaSettings.java
@@ -107,7 +107,7 @@ public final class CarpetAyakaSettings {
 
     @Rule(
             category = {AYAKA, EXPERIMENTAL, REINTRODUCE, FEATURE},
-            desc = "Reintroduces the feature of cactuses, bamboos and sugarcane being (wrongly) random-ticked on scheduled ticks in Minecraft 1.15.2 and lower"
+            desc = "Reintroduces the feature of cactuses, bamboos, chorus flowers and sugarcane being (wrongly) random-ticked on scheduled ticks in Minecraft 1.15.2 and lower"
     )
     @Condition(ForceTickPlantsReintroduceCondition.class)
     public static boolean forceTickPlantsReintroduce = false;