+++ changelog.md
@@ -1,3 +1,4 @@
 ## General
 
 - Added support for Minecraft 1.21.6
+- `forceTickPlantsReintroduce` now affects chorus flowers as well